Update the tan function to avoid asymptote and function overlap
Created restriction for x-values. Restricted the x-values for the tangent function to -270 to 271 and -180 to 181, there was still overlap. So, changed restriction to -90 to 91 and overlap is gone but now, there is only one period of the tangent function and multiple asymptotes. Will work on doing multiple tangent periods next.